Move a FormLayout component to the `layouts` directory
